recipe_creator/pyproject.toml
[project]
name = "rekipe-agent-recipe-creator"
version = "0.1.0"
requires-python = ">=3.13,<3.14"
dependencies = [
    "pydantic-ai>=1.69.0",
    "langgraph>=0.4.0",
    "langchain-mcp-adapters>=0.1.0",
    "chainlit>=2.10.0",
    "rapidfuzz>=3.0",
    "pyyaml>=6.0",
    "loguru>=0.7",
]

[tool.uv]
package = false


